feat: move `v0` schemas to `v0` module
