+++ src/argilla/server/apis/v0/handlers/datasets.py
@@ -25,7 +25,7 @@ from argilla.server.apis.v0.models.commons.params import (
 )
 from argilla.server.errors import EntityNotFoundError
 from argilla.server.models import User
-from argilla.server.schemas.datasets import CopyDatasetRequest, CreateDatasetRequest, Dataset, UpdateDatasetRequest
+from argilla.server.schemas.v0.datasets import CopyDatasetRequest, CreateDatasetRequest, Dataset, UpdateDatasetRequest
 from argilla.server.security import auth
 from argilla.server.services.datasets import DatasetsService
