+++ src/argilla/server/apis/v0/models/text_classification.py
@@ -26,7 +26,7 @@ from argilla.server.apis.v0.models.commons.model import (
     SortableField,
 )
 from argilla.server.commons.models import PredictionStatus
-from argilla.server.schemas.datasets import UpdateDatasetRequest
+from argilla.server.schemas.v0.datasets import UpdateDatasetRequest
 from argilla.server.services.search.model import ServiceBaseRecordsQuery, ServiceBaseSearchResultsAggregations
 from argilla.server.services.tasks.text_classification.model import (
     DatasetLabelingRulesMetricsSummary as _DatasetLabelingRulesMetricsSummary,
